requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.0

# LangChain and LLM
langchain>=0.1.0
langchain-community>=0.0.10

# Vector database - use version compatible with current Pydantic
chromadb<0.4.0  # 0.3.x supports Pydantic v1

# Embeddings
sentence-transformers>=2.2.0

# Web framework
streamlit>=1.28.0

# Data validation - keep v1 for chromadb compatibility
pydantic>=1.10,<2.0

# Configuration
pyyaml>=6.0

# Database - use compatible version
sqlalchemy>=1.4,<2.0

# Visualization
plotly>=5.17.0

# Scientific computing
scipy>=1.11.0

# Data storage
pyarrow>=12.0.0  # For parquet support

# Optional: For better performance
# numba>=0.58.0

# System monitoring
psutil>=5.9.0

